Compute sentiment for new terms.
I think this is compliant with the assignment, but there are a lot of
omissions from the problem statement:

  * what about nonsense words?
  * non-english words?
  * small words (articles)?

I don't see any mention of how to handle these on the assignment page
and I'm too lazy to look further.

Fun thing I did here was to subclass a staticmethod for the first time.
Using super() call was then interesting.

Assignment 1, problem 3.